Reverted prop handling to uninterned symbols because it broke a bunch of my code. Needs discussion:
Dash separated prop names. Given:

   (psx (:component :an-attrib 1))

We need to consider what happens to the prop name. As of this patch, it
is converted to camel case. The benefits are that retrieval from props
is quite natural: (prop an-attrib) works as expected, and anyone
interacting with the component from normal javascript will get the camel
case that they expect.

The downside is when we manually generate component props.

    > (ps (create :x 1 :an-attrib 2))
    "({ 'x' : 1, 'an-attrib' : 2 });"

The prop name needs a regular quote or to be uninterned:

    > (ps (create :x 1 'an-attrib 2))
    "({ 'x' : 1, 'anAttrib' : 2 });"

This misfeature has cost me an embarrassing amount of debugging time.
It would be nice to get the compiler to warn about it somehow.

The psx compiler determines that an item is a prop by checking if it is
a keyword, so we can't stick 'an-attrib in our psx elements. I don't
think we should change that. Readability is enhanced by the fact that
all of the props have leading colons.